Added background information.
This is the IP set program for the Arduino Ethernet Shield, it allows to set the IP and MAC address of the shield.

Code developed by Arduino.

The text of the Arduino reference is licensed under a Creative Commons Attribution-ShareAlike 3.0 License. Code samples in the reference are released into the public domain.